# riskcast-v16-main/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RISKCAST V3 API
  description: |
    Enterprise Insurance-Grade Risk Intelligence Platform API
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Versioning
    API is versioned via URL path (e.g., `/api/v3/...`)
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 1000 requests/minute
    
    ## Audit Trail
    All write operations are recorded in an immutable audit ledger with hash chain verification.
    
    ## Multi-Tenancy
    All operations are automatically scoped to the authenticated user's tenant.
    
    ## Permissions
    Endpoints require specific RBAC permissions. See individual endpoint documentation.
  version: 3.0.0
  contact:
    name: API Support
    email: api-support@riskcast.com
  license:
    name: Proprietary
    url: https://riskcast.com/terms
servers:
  - url: https://api.riskcast.com/api/v3
    description: Production
  - url: https://staging-api.riskcast.com/api/v3
    description: Staging
  - url: http://localhost:8000/api/v3
    description: Local Development
...
paths:
...
  /evidence/bundles:
    post:
      tags: [Evidence]
      summary: Create evidence bundle
      description: |
        Create a new evidence bundle.
        
        **Permission Required**: `EVIDENCE_WRITE`
      operationId: createEvidenceBundle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceBundleCreateRequest'
      responses:
        '201':
          description: Bundle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceBundleResponse'
  /evidence/bundles/{bundle_id}/seal:
    post:
      tags: [Evidence]
      summary: Seal evidence bundle
      description: |
        Seal the bundle, making it immutable.
        
        A manifest hash is computed from all items.
        
        **Permission Required**: `EVIDENCE_WRITE`
      operationId: sealEvidenceBundle
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BundleId'
      responses:
        '200':
          description: Bundle sealed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceBundleResponse'
  /evidence/bundles/{bundle_id}/verify:
    get:
      tags: [Evidence]
      summary: Verify bundle integrity
      description: |
        Verify that the evidence bundle has not been tampered with.
        
        **Permission Required**: `EVIDENCE_READ`
      operationId: verifyEvidenceBundle
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BundleId'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleVerificationResult'

  # ===========================================================================
  # Models
  # ===========================================================================
...
components:
...
  schemas:
...
    EvidenceBundleCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
              metadata:
                type: object
    EvidenceBundleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, SEALED]
        manifest_hash:
          type: string
        created_at:
          type: string
          format: date-time
    BundleVerificationResult:
      type: object
      properties:
        valid:
          type: boolean
        stored_hash:
          type: string
        computed_hash:
          type: string
        verified_at:
          type: string
          format: date-time
...
  parameters:
...
    BundleId:
      name: bundle_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Evidence bundle ID
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoint.
        Include in Authorization header as: `Bearer <token>`
security:
  - BearerAuth: []